+++ Tests/ReindexerNet.EmbeddedBenchmarks/BenchmarkHelper.cs
@@ -1,10 +1,6 @@
-﻿using ReindexerNet;
+﻿using LiteDB;
+using ReindexerNet;
 using ReindexerNetBenchmark.EmbeddedBenchmarks;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ReindexerNetBenchmark;
 
@@ -18,6 +14,9 @@ internal static class BenchmarkHelper
 
     public static object CaptureResult(this IEnumerable<BenchmarkEntity> entites)
         => entites.Select(e => e.PreventLazy()).ToList();
+
+    public static object CaptureResult(this ILiteQueryable<BenchmarkEntity> entites)
+        => entites.ToEnumerable().Select(e => e.PreventLazy()).ToList();
 
     public static object CaptureResult(this IEnumerable<BenchmarkRealmEntity> entites)
         => entites.Select(e => e.PreventLazy()).ToList();