+++ Tests/ReindexerNet.EmbeddedBenchmarks/SelectSingleHash.cs
@@ -65,17 +65,17 @@ public class SelectSingleHash : SelectBenchmarkBase
         return result;
     }
 
-    [Benchmark]
-    public IList<object?> CachalotCompressed()
-    {
-        var result = new List<object?>();
-        for (int i = 0; i < N; i++)
-        {
-            var str = Data[i].StringProperty;
-            result.Add(CaDSCompressed.FirstOrDefault(e => e.StringProperty == str).CaptureResult());
-        }
-        return result;
-    }
+    //[Benchmark]
+    //public IList<object?> CachalotCompressed()
+    //{
+    //    var result = new List<object?>();
+    //    for (int i = 0; i < N; i++)
+    //    {
+    //        var str = Data[i].StringProperty;
+    //        result.Add(CaDSCompressed.FirstOrDefault(e => e.StringProperty == str).CaptureResult());
+    //    }
+    //    return result;
+    //}
 
     [Benchmark]
     public IList<object?> LiteDb()